refactor: consolidate radar chart into shared GoalsRadarChart component
